Links can now show Register and Forgot modals
Anchor elements can now directly bring up the Register and Forgot
Password modal windows. Format for Register: <a href="#login-box"
class="login wpml-btn register-window">Register</a>
For Forgotten: <a href="#login-box" class="login wpml-btn
forgot-window">Register</a>

Added to wp-modal-login shortcode: Can now have parameter register_text.
If present, then the shortcode will be replaced with a Register link if
the user is NOT logged in and nothing if he is not.

Bug fix: when closing modal, reset the visibility of the divs after the
fade animation completes.